get quote working

+++ app/src/main/java/app/trial/eaglequotetrial/model/Premium.java
@@ -2,6 +2,6 @@ package app.trial.eaglequotetrial.model;
 
 public class Premium {
     public String policyFee;;
-    public long totalPremium;
+    public double totalPremium;
     public String errorMessage;
 }
