+++ app/src/main/java/app/trial/eaglequotetrial/model/ProviderPremium.java
@@ -1,7 +1,7 @@
 package app.trial.eaglequotetrial.model;
 
 public class ProviderPremium {
-    public long benefitTotalofPremiums;
-    public long policyFee;
-    public long totalPremium;
+    public double benefitTotalofPremiums;
+    public double policyFee;
+    public double totalPremium;
 }
