+++ app/src/main/java/app/trial/eaglequotetrial/model/Response.java
@@ -4,9 +4,9 @@ public class Response {
     public int code;
     public String status;
     public String message;
-    public long crunchTimeRequestFromQuoteEngine;
-    public long crunchTimeSummary;
-    public long crunchTimeSavingTrans;
+    public double crunchTimeRequestFromQuoteEngine;
+    public double crunchTimeSummary;
+    public double crunchTimeSavingTrans;
     public String email;
     public String password;
     public boolean rememberMe;
