+++ app/src/main/java/app/trial/eaglequotetrial/model/Result.java
@@ -2,7 +2,7 @@ package app.trial.eaglequotetrial.model;
 
 public class Result {
     public Provider[] providers;
-    public long crunchTimeQuoteEngine;
-    public long crunchTimeFidelity;
+    public double crunchTimeQuoteEngine;
+    public double crunchTimeFidelity;
     public String apiLastUpdated;
 }
